Refactor: relocate business logic from models
