+++ Quantra.DAL/Models/EmaCrossoverStrategy.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using Quantra.Utilities;
 
 namespace Quantra.Models
 {
