+++ Quantra.DAL/Models/SupportResistanceStrategy.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Windows.Media;
 using Quantra.DAL.Services.Interfaces;
+using Quantra.Utilities;
 
 namespace Quantra.Models
 {
